requirements.txt
pydantic>=2.0
pillow
tqdm
transformers
vllm
scikit-learn